split CSRF verification from the Authenticator
