+++ pkg/auth/oauth2/auth_test.go
@@ -49,7 +49,6 @@ func TestNewAuthenticator(t *testing.T) {
 		ErrorURL:      errURL,
 		SuccessURL:    sucURL,
 		CookiePath:    "/",
-		RefererPath:   "http://auth.example.com/",
 		SecureCookies: true,
 	}
 
@@ -97,7 +96,6 @@ func TestNewOpenShiftAuthenticator(t *testing.T) {
 		ErrorURL:      errURL,
 		SuccessURL:    sucURL,
 		CookiePath:    "/",
-		RefererPath:   "http://auth.example.com/",
 		SecureCookies: true,
 	}
 
@@ -138,7 +136,6 @@ func TestRedirectAuthError(t *testing.T) {
 		ErrorURL:      errURL,
 		SuccessURL:    sucURL,
 		CookiePath:    "/",
-		RefererPath:   "http://auth.example.com/",
 		SecureCookies: true,
 	}
 
@@ -166,8 +163,6 @@ func TestRedirectAuthError(t *testing.T) {
 	}
 }
 
-const validReferer string = "https://example.com/asdf/"
-
 func makeAuthenticator() (*OAuth2Authenticator, error) {
 	errURL := "https://example.com/error"
 	sucURL := "https://example.com/success"
@@ -179,7 +174,6 @@ func makeAuthenticator() (*OAuth2Authenticator, error) {
 		IssuerURL:     "http://auth.example.com",
 		ErrorURL:      errURL,
 		SuccessURL:    sucURL,
-		RefererPath:   validReferer,
 		SecureCookies: true,
 	}
 
@@ -189,107 +183,4 @@ func makeAuthenticator() (*OAuth2Authenticator, error) {
 	}
 
 	return newUnstartedAuthenticator(ccfg), nil
-}
-
-func testReferer(t *testing.T, referer string, accept bool) {
-	a, err := makeAuthenticator()
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	r, err := http.NewRequest("POST", "/some-path", nil)
-
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	if len(referer) > 0 {
-		r.Header.Set("Referer", referer)
-	}
-
-	err = a.VerifySourceOrigin(r)
-
-	if err != nil && accept {
-		t.Errorf("Unexpected error for referer `%v`:\n%v", referer, err)
-		return
-	}
-
-	if err == nil && !accept {
-		t.Errorf("Unexpected pass for referer: `%v:`\n%v", referer, err)
-	}
-
-	if accept {
-		t.Logf("referer accepted %v", referer)
-	} else {
-		t.Logf("referer rejected %v", referer)
-	}
-}
-
-func TestReferer(t *testing.T) {
-	testReferer(t, validReferer, true)
-	testReferer(t, validReferer, true)
-	testReferer(t, validReferer, true)
-	testReferer(t, validReferer+"other/path", true)
-	testReferer(t, validReferer+"?a=b&b=c#33", true)
-	testReferer(t, "", false)
-	testReferer(t, "http://example.com/asdf/", false)
-	testReferer(t, "http://example.com:8000/asdf/", false)
-	testReferer(t, "https://google.com/asdf/", false)
-	testReferer(t, "https://example.com/", false)
-	testReferer(t, "https://example.com/asdff/", false)
-	testReferer(t, "/asdff/", false)
-	testReferer(t, "🍆🍆🍆🍆🍆🍆", false)
-	testReferer(t, "https://google.com/asdf/", false)
-}
-
-func testCSRF(t *testing.T, token string, cookie string, accept bool) {
-	a, err := makeAuthenticator()
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	r, err := http.NewRequest("POST", "/some-path", nil)
-
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
-	if len(cookie) > 0 {
-		r.Header.Set(CSRFHeader, token)
-		r.AddCookie(&http.Cookie{
-			Name:     CSRFCookieName,
-			Value:    cookie,
-			MaxAge:   1000000,
-			HttpOnly: true,
-			Path:     "/",
-		})
-	}
-
-	err = a.VerifyCSRFToken(r)
-
-	if err != nil && accept {
-		t.Errorf("Unexpected error for CSRF `%v//%v`:\n%v", token, cookie, err)
-		return
-	}
-
-	if err == nil && !accept {
-		t.Errorf("Unexpected pass  for CSRF `%v//%v`:\n%v", token, cookie, err)
-	}
-
-	if accept {
-		t.Logf("CSRF accepted `%v` / `%v`", cookie, token)
-	} else {
-		t.Logf("CSRF rejected `%v` / `%v`", cookie, token)
-	}
-}
-func TestCSRF(t *testing.T) {
-	testCSRF(t, "a", "a", true)
-	testCSRF(t, "a", "b", false)
-	testCSRF(t, "a", "", false)
-	testCSRF(t, "", "b", false)
-	testCSRF(t, "", "", false)
 }
